Fix the issue of parameter type overflow in the governance mechanism

+++ src/TomorrowDAOServer.Application.Contracts/DAO/Dtos/DAOInfoDto.cs
@@ -74,11 +74,11 @@ public class FileDto
 
 public class GovernanceSchemeThresholdDto
 {
-    public int MinimalRequiredThreshold { get; set; }
-    public int MinimalVoteThreshold { get; set; }
-    public int MinimalApproveThreshold { get; set; }
-    public int MaximalRejectionThreshold { get; set; }
-    public int MaximalAbstentionThreshold { get; set; }
+    public long MinimalRequiredThreshold { get; set; }
+    public long MinimalVoteThreshold { get; set; }
+    public long MinimalApproveThreshold { get; set; }
+    public long MaximalRejectionThreshold { get; set; }
+    public long MaximalAbstentionThreshold { get; set; }
     public long ProposalThreshold { get; set; }
 }
